fix(core): don't ignore `ActionCancelled` exception from menu
In Delizia, it should be possible to cancel the flow using the menu.

[no changelog]
